// frontend/cinebook/src/app/services/models/booking-details.ts
/* tslint:disable */
/* eslint-disable */
export interface BookingDetails {
  bookingDate: string;
  bookingTime: string;
  movieLanguage: string;
  movieName: string;
  moviePosterUrl: string;
  numberOfTickets: number;
  seatNumbers: Array<string>;
  seatType: string;
  ticketId: string;
  username: string;
  ticketPrice: number;
  totalPrice: number;
}